requirements.txt
# Working Requirements - Tested and Verified
# Generated from working installation

# Core Web Framework
Flask==3.0.0
Flask-Cors==4.0.0
Werkzeug==3.1.5

# AI/ML Libraries
ultralytics==8.4.2
torch==2.9.1
torchvision==0.24.1

# Image Processing
opencv-python==4.12.0.88
Pillow==12.1.0
numpy==2.2.6

# Data Processing
PyYAML==6.0.3
requests==2.32.5

# Visualization
matplotlib==3.10.8

# System Utilities
psutil==7.2.1

# Production Server
gunicorn==21.2.0
